CZ: first attempt at exercises 1 and 2
added some functions to record build info, and produce an output file after running the test script. I'm trying to go in a direction that doesn't use any 3rd-party testing software, so we can underscore the basic principles of testing and reproducibility, saving efficiency for a later stage. More info can be found in the file "CZ-ex-attempt-1_noes.md"